// my-app/src/types/Profile.ts
export interface Project {
  title: string;
  description: string;
  link: string;
}

export interface Profile {
  name: string;
  email: string;
  college: string;
  university: string;
  degree: string;
  cgpa: string;
  skills: string[];
  achievements: string[];
  internshipsCompleted: number;
  appliedInternships: number;
  profileProgress: number;
}